Refactor status display functions to tui folder
- Move display_status and _display_usage_metrics_container to tui/status.py
- Update imports in agent_chat.py and tests
- Clean up unused imports in agent_chat.py
- Update test mocking to use new module paths
- All 8 unit tests still pass
- Better organization with UI components in tui folder
